Move fixit.types to fixit.ftypes
This helps prevent obnoxious warnings from vscode/pylance about
shadowing the stdlib `types` module.

[ghstack-poisoned]
