+++ src/fixit/__init__.py
@@ -10,10 +10,10 @@ Linting framework built on LibCST, with automatic fixes
 __version__ = "0.2.0"
 
 from .api import fixit_bytes, fixit_file, fixit_paths
+from .ftypes import CodePosition, CodeRange, Config, FileContent, Result
 from .rule import LintRule
 from .rule.cst import CSTLintRule, CstLintRule
 from .testing import InvalidTestCase, ValidTestCase
-from .types import CodePosition, CodeRange, Config, FileContent, Result
 
 __all__ = [
     "__version__",
